% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_fjlt}
\alias{calc_fjlt}
\title{Calculate the FJLT transform}
\usage{
calc_fjlt(x, p, k, d, n)
}
\arguments{
\item{p}{the norm}

\item{k}{dimension we reduce to}

\item{d}{dimension of the input matrix}

\item{n}{number of cells (columns) in the input matrix}

\item{epsilon}{error tolerance parameter}
}
\value{
FJLT transform
}
\description{
Fast-Johnson-Lindenstrauss-Transform
}

