% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateFJLT}
\alias{calculateFJLT}
\title{Calculate the FJLT transform}
\usage{
calculateFJLT(x, p, k, d, n)
}
\arguments{
\item{p}{the norm}

\item{k}{dimension we reduce to}

\item{d}{dimension of the input matrix}

\item{n}{number of cells (columns) in the input matrix}

\item{epsilon}{error tolerance parameter}
}
\value{
FJLT transform
}
\description{
Fast-Johnson-Lindenstrauss-Transform
}

