% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{constr_D}
\alias{constr_D}
\title{Construct the D matrix for the FJLT transform}
\usage{
constr_D(d)
}
\arguments{
\item{d}{dimension of the input matrix}
}
\value{
matrix D of the FJLT transform
}
\description{
Construct the D matrix for the FJLT transform
}

