% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{constr_H}
\alias{constr_H}
\title{Construct the H (Hadamard) matrix for the FJLT transform}
\usage{
constr_H(d)
}
\arguments{
\item{d}{dimension of the input matrix}
}
\value{
matrix H of the FJLT transform
}
\description{
Construct the H (Hadamard) matrix for the FJLT transform
}

