% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{constructH}
\alias{constructH}
\title{Construct the H (Hadamard) matrix for the FJLT transform}
\usage{
constructH(d)
}
\arguments{
\item{d}{dimension of the input matrix}
}
\value{
matrix H of the FJLT transform
}
\description{
Construct the H (Hadamard) matrix for the FJLT transform
}

