% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{constructP}
\alias{constructP}
\title{Construct the P matrix for the FJLT transform}
\usage{
constructP(p, k, d, n)
}
\arguments{
\item{p}{the norm}

\item{k}{dimension we reduce to}

\item{d}{dimension of the input matrix}

\item{n}{number of cells (columns) in the input matrix}

\item{epsilon}{error tolerance parameter}
}
\value{
matrix P of the FJLT transform
}
\description{
Construct the P matrix for the FJLT transform
}

