% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{FJLT}
\alias{FJLT}
\title{Fast-Johnson-Lindenstrauss-Transform (FJLT)}
\usage{
FJLT(x, k = 100)
}
\arguments{
\item{x}{input expression matrix}

\item{epsilon}{error tolerance parameter}
}
\value{
transformed reduced expression matrix
}
\description{
This function calculates the FJLT and project onto d dimensions. The FJLT 
is faster than standard random pro-jections and just as easy to implement. 
It is based upon the preconditioning of a sparse projection matrix with a 
randomized Fourier transform.
}
\details{
Ailon, N. & Chazelle, B. Approximate nearest neighbors and the fast 
Johnson-Lindenstrauss transform. in Proceedings of the thirty-eighth 
annual ACM symposium on Theory of computing 557–563 (ACM, 2006).

Functions adapted from: http://www.cs.ubc.ca/~jaquesn/MachineLearningTheory.pdf
}

