% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{fsc3_calc_eigenv.SCESet}
\alias{fsc3_calc_eigenv}
\alias{fsc3_calc_eigenv,SCESet-method}
\alias{fsc3_calc_eigenv.SCESet}
\title{Calculate eigenvectors of the transformed distance matrix.}
\usage{
fsc3_calc_eigenv.SCESet(object)

\S4method{fsc3_calc_eigenv}{SCESet}(object)
}
\arguments{
\item{object}{an object of 'SCESet' class}
}
\value{
an object of 'SCESet' class
}
\description{
This function calculates eigenvectors of the transformed distance matrices contained in 
the transformation item of the object@sc3 slot. It then
populates back the transformations item the calculated eigenvectors.
}
\details{
Eigenvectors are calculated by using the SS-Nystrom approximation.
Wang, S., Zhang, C., Qian, H. & Zhang, Z. Improving the Modified NyströM 
Method Using Spectral Shifting. in Proceedings of the 20th ACM SIGKDD 
International Conference on Knowledge Discovery and Data Mining 611–620 
(ACM, 2014).
}

