% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{fsc3_fjlt.SCESet}
\alias{fsc3_fjlt}
\alias{fsc3_fjlt,SCESet-method}
\alias{fsc3_fjlt.SCESet}
\title{Fast-Johnson-Lindenstrauss-Transform (FJLT)}
\usage{
fsc3_fjlt.SCESet(object, k = 100)

\S4method{fsc3_fjlt}{SCESet}(object, k = 100)
}
\arguments{
\item{object}{an object of 'SCESet' class}

\item{k}{number of dimensions to reduce to}
}
\description{
This function calculates the FJLT and project onto d dimensions. The FJLT 
is faster than standard random projections and just as easy to implement. 
It is based upon the preconditioning of a sparse projection matrix with a 
randomized Fourier transform.
}
\details{
Ailon, N. & Chazelle, B. Approximate nearest neighbors and the fast 
Johnson-Lindenstrauss transform. in Proceedings of the thirty-eighth 
annual ACM symposium on Theory of computing 557–563 (ACM, 2006).

Calculate Fast-Johnson-Lindenstrauss-Transform (FJLT)

This function calculates FJLT projection of
the processed_dataset item of the object@sc3 slot. It then
creates and populates the following items of the object@sc3 slot:
\itemize{
  \item processed_dataset_fjlt - contains the FJLT transform of the processed_dataset
}
}

