% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking.R
\name{fsc3_get_buckets_med_exprs}
\alias{fsc3_get_buckets_med_exprs}
\title{Calculate median feature expression in every bucket}
\usage{
fsc3_get_buckets_med_exprs(object)
}
\arguments{
\item{object}{an object of 'SCESet' class}
}
\value{
an object of 'SCESet' class
}
\description{
When the cell buckets are defined, it is possible to calculate a median
expression based on the expressions of the cells contained in a given
bucket.
}

