% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{fsc3_get_buckets_signatures.SCESet}
\alias{fsc3_get_buckets_signatures}
\alias{fsc3_get_buckets_signatures,SCESet-method}
\alias{fsc3_get_buckets_signatures.SCESet}
\title{Calculate average bucket signatures}
\usage{
fsc3_get_buckets_signatures.SCESet(object)

\S4method{fsc3_get_buckets_signatures}{SCESet}(object)
}
\arguments{
\item{object}{an object of 'SCESet' class}
}
\value{
an object of 'SCESet' class
}
\description{
When the cell buckets are defined, it is possible to calculate an average
bucket signature based on the signatures of the cells contained in a given
bucket.
}

