% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{fsc3_get_clusters.SCESet}
\alias{fsc3_get_clusters}
\alias{fsc3_get_clusters,SCESet-method}
\alias{fsc3_get_clusters.SCESet}
\title{Reindex cell buckets based on the merged buckets results}
\usage{
fsc3_get_clusters.SCESet(object, k)

\S4method{fsc3_get_clusters}{SCESet}(object, k)
}
\arguments{
\item{object}{an object of 'SCESet' class}

\item{k}{number of clusters}
}
\value{
an object of 'SCESet' class
}
\description{
Reindex cell buckets based on the merged buckets results
}

