% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{fsc3_get_features.SCESet}
\alias{fsc3_get_features}
\alias{fsc3_get_features,SCESet-method}
\alias{fsc3_get_features.SCESet}
\title{Define a set of genes used for creating cell binary signatures}
\usage{
fsc3_get_features.SCESet(object, n_genes = 100, pct_dropout_min = 20,
  pct_dropout_max = 80, suppress_plot = T)

\S4method{fsc3_get_features}{SCESet}(object, n_genes = 100,
  pct_dropout_min = 20, pct_dropout_max = 80, suppress_plot = T)

fsc3_get_features.SCESet(object, n_genes = 100, pct_dropout_min = 20,
  pct_dropout_max = 80, suppress_plot = T)

\S4method{fsc3_get_features}{SCESet}(object, n_genes = 100,
  pct_dropout_min = 20, pct_dropout_max = 80, suppress_plot = T)
}
\arguments{
\item{object}{an object of 'SCESet' class}

\item{n_genes}{number of the genes to be returned}

\item{object}{an object of 'SCESet' class}

\item{n_genes}{number of the genes to be returned}
}
\value{
an object of 'SCESet' class

an object of 'SCESet' class
}
\description{
The important genes (hyperplanes) are defined as differentially expressed 
genes identified using M3Drop (Michaelis-Menten Modelling of Dropouts for scRNASeq,
http://bioconductor.org/packages/M3Drop). A 'data.frame' with the genes
and corresponding p and q values is written to the 'hyperplanes' item of the
'sc3' slot of the input object.

The important genes (hyperplanes) are defined as differentially expressed 
genes identified using M3Drop (Michaelis-Menten Modelling of Dropouts for scRNASeq,
http://bioconductor.org/packages/M3Drop). A 'data.frame' with the genes
and corresponding p and q values is written to the 'hyperplanes' item of the
'sc3' slot of the input object.
}

