% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{fsc3_get_signatures.SCESet}
\alias{fsc3_get_signatures}
\alias{fsc3_get_signatures,SCESet-method}
\alias{fsc3_get_signatures.SCESet}
\title{Create a binary signature for each cell}
\usage{
fsc3_get_signatures.SCESet(object)

\S4method{fsc3_get_signatures}{SCESet}(object)
}
\arguments{
\item{object}{an object of 'SCESet' class}
}
\value{
an object of 'SCESet' class
}
\description{
Create a binary signature for each cell
}

