% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{fsc3_norm_kernel.SCESet}
\alias{fsc3_norm_kernel}
\alias{fsc3_norm_kernel,SCESet-method}
\alias{fsc3_norm_kernel.SCESet}
\title{Calculate doubly-stochastic approximation of the PSD matrix}
\usage{
fsc3_norm_kernel.SCESet(object)

\S4method{fsc3_norm_kernel}{SCESet}(object)
}
\arguments{
\item{object}{an object of 'SCESet' class}
}
\value{
an object of 'SCESet' class
}
\description{
Zass, R. and Shashua, A. in Advances in Neural Information Processing 
Systems 19 (eds. Scholkopf, B., Platt, J. C. and Hoffman, T.) 1569-1576 
(MIT Press, 2007).
}

