% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{fsc3_set_features.SCESet}
\alias{fsc3_set_features}
\alias{fsc3_set_features,SCESet-method}
\alias{fsc3_set_features.SCESet}
\title{Title}
\usage{
fsc3_set_features.SCESet(object, features)

\S4method{fsc3_set_features}{SCESet}(object, features)
}
\arguments{
\item{features}{a vector of feature names}
}
\description{
Title
}

