% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{ftransformation}
\alias{ftransformation}
\title{Distance matrix transformation}
\usage{
ftransformation(dists, method)
}
\arguments{
\item{dists}{distance matrix}

\item{method}{transformation method: either 'pca' or
'laplacian'}
}
\value{
transformed distance matrix
}
\description{
All distance matrices are transformed to either covariance matrix 
(in case of PCA) or to graph Laplacian (in case of Laplacian).
}

