% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pad_matrix_with_zeros}
\alias{pad_matrix_with_zeros}
\title{Calculate the FJLT transform}
\usage{
pad_matrix_with_zeros(x, d)
}
\arguments{
\item{x}{input matrix}

\item{d}{dimension to pad to}
}
\value{
FJLT transform
}
\description{
Fast-Johnson-Lindenstrauss-Transform
}

