% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ssNystrom}
\alias{ssNystrom}
\title{Calculate approximate eigenvectors}
\usage{
ssNystrom(K, r)
}
\arguments{
\item{K}{input SPSD matrix}

\item{p}{number of selected columns as fraction of ncol(K)}
}
\value{
approximate eigenvectors
}
\description{
Nystrom spectral shifting method
}

