% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{fjlt}
\alias{fjlt}
\title{Fast-Johnson-Lindenstrauss-Transform (FJLT)}
\usage{
fjlt(x, k = 100)
}
\arguments{
\item{x}{input expression matrix}

\item{k}{number of dimension to reduce to}
}
\value{
transformed reduced expression matrix
}
\description{
This function calculates the FJLT and project onto d dimensions. The FJLT 
is faster than standard random pro-jections and just as easy to implement. 
It is based upon the preconditioning of a sparse projection matrix with a 
randomized Fourier transform.
}
\details{
Ailon, N. and Chazelle, B. Approximate nearest neighbors and the fast 
Johnson-Lindenstrauss transform. in Proceedings of the thirty-eighth 
annual ACM symposium on Theory of computing 557–563 (ACM, 2006).

Functions adapted from: \url{http://www.cs.ubc.ca/~jaquesn/MachineLearningTheory.pdf}
}

